Import code

+++ run_tests.py
@@ -7,9 +7,9 @@ from multiprocessing import Pool
 import argparse
 
 
-ROOT_DIR = os.getcwd()
 DEFAULT_MP_POOL_SIZE = 128
 DEFAULT_NUM_SEEDS = 100
+ROOT_DIR = os.getcwd()
 
 gjrand_exe = os.path.join(ROOT_DIR, 'install', 'gjrand.4.3.0.0', 'testunif', 'pmcp')
 gjrand_bin = os.path.join(ROOT_DIR, 'install', 'gjrand.4.3.0.0', 'testunif', 'bin')
@@ -17,17 +17,18 @@ gjrand_bin = os.path.join(ROOT_DIR, 'install', 'gjrand.4.3.0.0', 'testunif', 'bi
 # Setup paths to generator and test binaries
 custom_env = os.environ.copy()
 custom_env["PATH"] = ROOT_DIR + ':' + custom_env["PATH"]
-custom_env["PATH"] = os.path.join(ROOT_DIR, 'bin') + ':' + custom_env["PATH"]
+custom_env["PATH"] = os.path.join(ROOT_DIR, 'build') + ':' + custom_env["PATH"]
 custom_env["PATH"] = os.path.join(ROOT_DIR, 'install', 'PractRand-pre0.95') + ':' + custom_env["PATH"]
 
 
 class TestU01Arguments:
-    def __init__(self, generator, testsuite, output):
+    def __init__(self, generator, testsuite, output, output_shift):
         self.generator = generator
         self.testsuite = testsuite
         self.output = output
+        self.output_shift = output_shift
     def get_args(self, s0, s1):
-        return [self.generator, self.testsuite, self.output, s0, s1]
+        return [self.generator, self.testsuite, self.output, s0, s1, '--output-shift', str(self.output_shift)]
 
 class CmdLineArguments:
     def __init__(self, generator, output, extra_args):
@@ -35,16 +36,15 @@ class CmdLineArguments:
         self.output = output
         self.extra_args = extra_args
     def get_args(self, s0, s1):
-        return ['bash', '-c', f'"{self.generator} stdout {self.output} {s0} {s1} | {extra_args}"']
+        return ['bash', '-c', f'"{self.generator} stdout {self.output} {s0} {s1} | {self.extra_args}"']
 
 
-def run_on_queue(cmd, working_directory, dry_run):
+def run_command(cmd, working_directory, dry_run):
     name = working_directory.replace('/', '_')
-    qsub = f'qsub -q <queue> -V -sync y -j y -cwd -b y -l num_proc=1 -N {name}'.split()
-    print("Running: '{}' in {}".format((' '.join(qsub + cmd)), working_directory))
+    print("Running: '{}' in {}".format((' '.join(cmd)), working_directory))
     try:
         if not dry_run:
-            subprocess.check_call(qsub + cmd, cwd=working_directory, env=custom_env)
+            subprocess.check_call(cmd, cwd=working_directory, env=custom_env)
     except subprocess.CalledProcessError as e:
         print(e.output)
 
@@ -65,19 +65,22 @@ def run_test(index, testsuite, num_seeds, gen_args, output_directory, dry_run):
         shutil.copy(gjrand_exe, os.path.join(working_directory, 'pmcp'))
         shutil.copytree(gjrand_bin, os.path.join(working_directory, 'bin'))
     # Run the job.
-    run_on_queue(gen_args.get_args(str(seed_lo64), str(seed_hi64)), working_directory, dry_run)
+    run_command(gen_args.get_args(str(seed_lo64), str(seed_hi64)), working_directory, dry_run)
 
 
-def run_all_seeds(generator, testsuite, output, num_seeds, output_directory, dry_run=False, mppoolsize=DEFAULT_MP_POOL_SIZE):
+def run_all_seeds(generator, testsuite, output, num_seeds, output_shift, output_directory, dry_run=False, mppoolsize=DEFAULT_MP_POOL_SIZE):
     # Setup the output directory.
-    if os.path.exists(output_directory):
+    if os.path.exists(output_directory) and not dry_run:
         print('Removing old results directory: ' + output_directory)
         shutil.rmtree(output_directory)
-    if not os.path.exists(output_directory):
+    if not os.path.exists(output_directory) and not dry_run:
         os.makedirs(output_directory)
     # Setup the command to run.
-    if testsuite.endswith('crush'):
-        gen_args = TestU01Arguments(generator, testsuite, output)
+    if testsuite.endswith('crush') or \
+       testsuite == 'alphabit' or \
+       testsuite == 'matrixrank' or \
+       testsuite == 'linearcomp':
+        gen_args = TestU01Arguments(generator, testsuite, output, output_shift)
     if testsuite == 'practrand' or \
        testsuite == 'gjrand':
         if testsuite == 'practrand':
@@ -89,9 +92,9 @@ def run_all_seeds(generator, testsuite, output, num_seeds, output_directory, dry
             # executables in a 'bin' folder. To work around this, the pmcp test
             # program and bin folder are copied and run from the working
             # directory.
-            tester_args = './pmcp --standard'
-            #tester_args = './pmcp --ten-tera'
-        gen_args = CmdLineArgs(generator, output, tester_args)
+            #tester_args = './pmcp --standard'
+            tester_args = './pmcp --ten-tera'
+        gen_args = CmdLineArguments(generator, output, tester_args)
     # Run jobs in parallel
     with Pool(mppoolsize) as p:
         p.starmap(run_test, zip(list(range(num_seeds)),
@@ -109,7 +112,7 @@ def main():
     parser.add_argument('generator',
                         help='Specify a path to the generator binary')
     parser.add_argument('testsuite',
-                        choices=['smallcrush', 'crush', 'bigcrush', 'practrand', 'gjrand'],
+                        choices=['smallcrush', 'crush', 'bigcrush', 'alphabit', 'practrand', 'gjrand', 'matrixrank', 'linearcomp'],
                         help='Specify a test suite to run')
     parser.add_argument('output',
                         help='Specify a generator output')
@@ -117,6 +120,10 @@ def main():
                         type=int,
                         default=DEFAULT_NUM_SEEDS,
                         help='The number of seeds to test')
+    parser.add_argument('--outputshift',
+                        type=int,
+                        default=0,
+                        help='The number of bits to right-shift the output by (only for matrixrank and binarycomp)')
     parser.add_argument('--dryrun',
                         action='store_true',
                         help='Don\'t run anything')
@@ -135,6 +142,7 @@ def main():
                   args.testsuite,
                   args.output,
                   args.numseeds,
+                  args.outputshift,
                   args.outputdir,
                   args.dryrun,
                   args.mppoolsize)
